refactor: 코스 및 지도 관련 CRUD 커스텀훅 정리
