+++ src/components/common/CourseMemo.tsx
@@ -1,5 +1,5 @@
 import TextareaAutosize from "react-textarea-autosize";
-import { useAddMemo, useCourse, useHandleMemo } from "../../hooks";
+import { useCourse } from "../../hooks";
 import { Dispatch, SetStateAction } from "react";
 
 interface CourseInfoProps {
@@ -11,8 +11,8 @@ interface CourseInfoProps {
 
 const CourseMemo = ({ idx, item, text, setText }: CourseInfoProps) => {
   const { filteredIdx } = useCourse();
-  const handleMemo = useHandleMemo();
-  const addMemo = useAddMemo();
+  const { handleMemo } = useCourse();
+  const { addMemo } = useCourse();
 
   return (
     <TextareaAutosize
