+++ src/components/common/SearchModal.tsx
@@ -1,4 +1,4 @@
-import { Dispatch, SetStateAction, useState } from "react";
+import React, { Dispatch, SetStateAction, useState } from "react";
 import { GrFormClose } from "react-icons/gr";
 import { useDispatch } from "react-redux";
 import { addCourse } from "../../redux/modules/courseSlice";
@@ -22,6 +22,8 @@ const SearchModal = ({
   setSearchList,
   searchCnt,
 }: PostProps) => {
+  console.log("렌더링 발생");
+
   const dispatch = useDispatch();
   const [text, setText] = useState("");
   const { lists } = useCourse();
@@ -137,4 +139,4 @@ const SearchModal = ({
   );
 };
 
-export default SearchModal;
+export default React.memo(SearchModal);
