+++ src/hooks/index.ts
@@ -2,12 +2,6 @@ import usePagenation from "./usePagenation";
 import usePreventLeave from "./usePreventLeave";
 import { useScript } from "./useScript";
 import useKakaoMap from "./useKakaoMap";
-import useFilterCourse from "./useFilterCourse";
-import useUpCourse from "./useUpCourse";
-import useDownCourse from "./useDownCourse";
-import useDeleteCourse from "./useDeleteCourse";
-import useHandleMemo from "./useHandleMemo";
-import useAddMemo from "./useAddMemo";
 import useBefore from "./useBefore";
 import useCommentPost from "./useCommentPost";
 import useDelete from "./useDelete";
@@ -22,12 +16,6 @@ export {
   usePreventLeave,
   useScript,
   useKakaoMap,
-  useFilterCourse,
-  useUpCourse,
-  useDownCourse,
-  useDeleteCourse,
-  useHandleMemo,
-  useAddMemo,
   useBefore,
   useCommentPost,
   useDelete,
