--- src/hooks/useUpCourse.ts
@@ -1,14 +1,0 @@
-import { useDispatch } from "react-redux";
-import { upCourse } from "../redux/modules/courseSlice";
-
-const useUpCourse = () => {
-  const dispatch = useDispatch();
-
-  const liftUp = (idx: number) => {
-    dispatch(upCourse(idx));
-  };
-
-  return liftUp;
-};
-
-export default useUpCourse;
